third_party/WebKit/LayoutTests/platform/win/http/tests/fetch/workers/thorough/scheme-blob-base-https-other-https-expected.txt

CONSOLE WARNING: line 14: Subresource requests whose URLs contain embedded credentials (e.g. `https://user:pass@host/`) are deprecated, and will be blocked in M59, around June 2017. See https://www.chromestatus.com/feature/5669008342777856 for more details.
CONSOLE ERROR: Only 'GET' method is allowed for blob URLs.
CONSOLE ERROR: Only 'GET' method is allowed for blob URLs.
This is a testharness.js-based test.
PASS Startup 
PASS executeTest-0 
PASS executeTest-1 
PASS executeTest-2 
PASS executeTest-3 
PASS executeTest-4 
Harness: the test ran to completion.
